Add QR scan view for restaurant management; integrate reservation approval logic and update navigation
