+++ lib/src/app.dart
@@ -3,6 +3,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:table_order/src/views/auth_view/login_page_view.dart';
 import 'package:table_order/src/views/navigation_rail_page.dart';
+import 'package:table_order/src/views/owner_view/restaurant_management_view/qr_scan_view.dart';
 import 'package:table_order/src/views/owner_view/restaurant_owner_page_view.dart';
 import 'package:table_order/src/views/owner_view/restaurant_registration.dart';
 import 'package:table_order/src/views/qr_view/reservation_qr_view.dart';
@@ -90,6 +91,8 @@ class MyApp extends StatelessWidget {
                     return const ReservationQrView();
                   case ReservationItemListView.routeName:
                     return const ReservationItemListView();
+                  case QrScanView.routeName:
+                    return const QrScanView();
                   case RestaurantItemListView.routeName:
                   default:
                     return const NavigationRailPage();
